requirements.txt
scikit-image
robotframework-browser
opencv-python
numpy
pytesseract
Pillow